Fixed overflow issue that would only happen on 64 bit compiles with hash tables with more than 2^31 buckets.

+++ src/mathic/HashTable.h
@@ -145,8 +145,8 @@ namespace mathic {
   template<class C>
   HashTable<C>::HashTable(const Configuration &conf, unsigned int nbits):
     mLogTableSize(nbits),
-    mTableSize(1 << nbits),
-    mHashMask((1 << nbits) - 1),
+    mTableSize(static_cast<size_t>(1) << nbits),
+    mHashMask((static_cast<size_t>(1) << nbits) - 1),
     mNodeCount(0),
     mBinCount(0),
     mRebuildThreshold(0.1),
